Fix nonce-based transport sequencing and recover on client address changes.
Use explicit Noise nonces with ROC-style reconstruction to prevent replay/decrypt desync after seq wrap, and reset stale SSRC sessions when the client's UDP source address changes so re-handshakes succeed after restarts/NAT rebinding.

Made-with: Cursor

+++ phantom-vpn/crates/core/src/error.rs
@@ -21,7 +21,7 @@ pub enum PacketError {
     #[error("Invalid SSRC: {0:#010x}")]
     InvalidSsrc(u32),
     #[error("Replay detected: seq={0}")]
-    Replay(u16),
+    Replay(u64),
     #[error("Packet timestamp too old: diff={0}s")]
     StaleTimestamp(u32),
     #[error("Decryption failed")]
